requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pandas==2.2.3
numpy==1.26.4
scipy==1.13.0
scikit-learn==1.4.2
xgboost==2.0.3
plotly==5.20.0
openpyxl==3.1.2
google-generativeai==0.8.3
google-genai==0.2.2
python-dotenv==1.0.1
werkzeug==3.0.1
gunicorn==21.2.0

